fixed ased_interact to accommodate interaction with gpt2 - add special token

+++ generate_valid.py
@@ -112,8 +112,8 @@ def run():
     parser.add_argument("--max_length", type=int, default=50, help="Maximum length of the output utterances")
     parser.add_argument("--min_length", type=int, default=1, help="Minimum length of the output utterances")
     parser.add_argument("--seed", type=int, default=42, help="Seed")
-    parser.add_argument("--temperature", type=int, default=0.7, help="Sampling softmax temperature")
-    parser.add_argument("--top_k", type=int, default=0, help="Filter top-k tokens before sampling (<=0: no filtering)")
+    parser.add_argument("--temperature", type=float, default=0.7, help="Sampling softmax temperature")
+    parser.add_argument("--top_k", type=float, default=0, help="Filter top-k tokens before sampling (<=0: no filtering)")
     parser.add_argument("--top_p", type=float, default=0.9, help="Nucleus filtering (top-p) before sampling (<=0.0: no filtering)")
     # add option to not use personality
     parser.add_argument("--no_personality", type=bool, default=True, help="Set to not sample a personality.")
